be stricter in rock_prepare_pkgconfig
Loadable modules don't need a pkg-config file (actually, the
file is basically useless). In the end, only rock_library needs
to generate the file..

This commit moves the call to rock_prepare_pkgconfig to rock_library.
It also makes the call strict (you have to mean it if you call it),
and promotes the warning about a missing pkg-config file (now in
rock_library) to WARNING