split rock_vizkit_widget between module and shared lib
"Recent" cmake (post-2.2) will refuse to link to a module. Since
MacOSX requires to make the distinction between modules and
shared libs, we have to split the Qt plugins between the module
part and the shared lib part.

This introduces a backward compatible mode, where we basically
build the shared lib twice. However, if the rock_vizkit_widget
stanza is updated to use the new QTPLUGIN <sourcefile> flag, it
then links the module to the shared lib, only adding the qtplugin
file to the module.

To make this work, I had to start installing the vizkit3d plugins
from lib/ to lib/vizkit3d - to avoid name clashes between the
pure shared library and the plugin