allow to be strict w.r.t. error handling in bindings/ruby/CMakeLists.txt
Right now, there is no way to know whether the user wants the ruby bindings
or not. They are basically automagically ignored if the "right" packages
cannot be found.

This commit changes that by making their inclusion explicit. By default,
they are included and an error should be generated if dependencies
are not available (e.g. rice). The user must set BINDINGS_RUBY to OFF
to ignore the bindings.

Note that while the rock_ruby_rice_extension macro would gracefully
ignore an extension if rice is not available, rock_ruby_extension would
not (leading to a compilation error). Moreover, the error handling
in FindGem was broken, which was leading to having compilation errors
even with rock_ruby_rice_extension if rice was not available.

In other words, compilation nowadays would have failed if rice was
not present. This commit only simplifies the matter by allowing to
be strict and generate proper errors in the cmake code, instead of
having weird compilation issues.