fix: do not throw fatal error when lacking the config path option for clang-format
Unfortunately, clang-format only allows setting an explicit config
path for clang-format-14 or later. So instead of throwing a fatal error
when it is unset, it will default to the "-style=file" string and if it
is set it will change to the version with an explicit path. For now,
it will be up to the user to fix their own setup when using a version
that doesnt support this option.